requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
alembic==1.15.2
PyMuPDF==1.25.5
python-docx==1.1.2
ollama==0.4.7
reportlab==4.2.5
